requirements/requirements-dev.txt
aspy.yaml==1.3.0
astroid==2.0.4
atomicwrites==1.3.0
attrs==19.3.0
aws-xray-sdk==0.95
azure-common==1.1.23
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto==2.49.0
boto3==1.7.73
botocore==1.10.84
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
cfgv==2.0.1
chardet==3.0.4
cookies==2.2.1
cryptography==2.8
distro==1.4.0
docker==4.1.0
docutils==0.15.2
dodgy==0.1.9
ecdsa==0.13.3
fake-factory==0.7.2
future==0.18.1
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-compute-engine==2.8.3
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
httpretty==0.8.14
identify==1.4.7
idna==2.8
importlib-metadata==0.23
isort==4.3.21
Jinja2==2.10.3
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
moto==1.3.4
nodeenv==1.3.3
packaging==19.2
pep8-naming==0.4.1
pluggy==0.13.0
pre-commit==1.16.0
prospector==1.1.6.2
protobuf==3.10.0
py==1.8.0
pyaml==19.4.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.4.0
pycparser==2.19
pycryptodome==3.9.0
pydocstyle==4.0.1
pyflakes==1.6.0
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.6
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.0
python-decouple==3.1
python-jose==2.0.2
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requirements-detector==0.6
responses==0.10.6
rsa==4.0
s3transfer==0.1.13
setoptconf==0.2.0
six==1.12.0
snowballstemmer==2.0.0
toml==0.10.0
urllib3==1.25.6
virtualenv==16.7.7
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
